Updates to reposition and align ratings
Adjustments and minor cleanup to reposition the internet ratings images/text up to the same line as the title, with some pixel value recalcs to ensure everything is again aligned in conjunction with the previous font definition fix.

Shifted the control group for the ratings up in the code so it is next to the group for the other items on line one (title, edition, etc.).